% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmp_DirLapl_mcmc.R
\name{nmp_DirLapl_mcmc}
\alias{nmp_DirLapl_mcmc}
\title{Normal Means Problem with Dirichlet-Laplace Prior: Markov Chain Monte
Carlo Algorithm}
\usage{
nmp_DirLapl_mcmc = function(nsim, y, alpha, burn = 0, thin = 1,
                                   verbose = +Inf)
}
\arguments{
\item{nsim}{The number of simulations.}

\item{y}{The response vector.}

\item{alpha}{The hyper-parameter controlling the amount of shrinkage.}

\item{burn}{The number of draws to be discarded as burn-in.}

\item{thin}{The thinning parameter.}

\item{verbose}{The period for printing status of the chain.}
}
\value{
Posterior sample of the parameters.
}
\description{
Gibbs sampler algorithm for the normal means problem with a
Dirichlet-Laplace prior.
}
\details{
The normal means problem is a statistical model in which a single
vector \eqn{y} is observed and is modeled as independent normal random
variables with known unit variance, say
\deqn{
 y \sim N_n (\theta, I_n)
}
where \eqn{I_n} is the \eqn{n}-dimensional identity matrix. The prior on the
mean vector \eqn{\theta} is the Dirichlet-Laplace prior:
\deqn{
 \theta_i \overset{ind}{\sim} N(0, \delta^2_i \psi_i) \, , \, \delta_i
  \overset{i.i.d.}{\sim} \mathrm{gamma}(\alpha, 1 / 2) \perp\!\!\!\perp
  \psi_i \overset{i.i.d.}{\sim} \exp(1 / 2)
 }
or, equivalently
\deqn{
  \theta_i \overset{ind}{\sim} N(0, \tau^2 \phi^2_i \psi_i) \\
  \tau \sim \mathrm{gamma}(n \alpha, 1 / 2) \\
  \phi \sim \mathrm{Dirichlet}_n(\alpha, \alpha, \dots, \alpha) \\
  \psi_i \overset{i.i.d.}{\sim} \exp(1 / 2)
 }
For the Gibbs sampler, it is used the first representation.

Only one value every \code{thin} values is kept in the chain, so the true
number of complete scans will be \code{nsim * thin + burn}. By default
\code{thin = 1}, that is no thinning.

The current time and the current number of iteration are printed one every
\code{verbose} iterations. Furthermore:
\itemize{
\item if \code{verbose == +-Inf} then there is no printing,
\item if \code{verbose != +-Inf} then at least start and end of simulation
are reported.
}
}
